# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"ExamMemory AI","version":"0.1.0"},"paths":{"/api/health":{"get":{"summary":"Health","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/users":{"post":{"summary":"Create User","operationId":"create_user_api_users_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/users/{user_id}":{"get":{"summary":"Get User","operationId":"get_user_api_users__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/articles":{"get":{"summary":"List Articles","operationId":"list_articles_api_articles_get","parameters":[{"name":"exam","in":"query","required":false,"schema":{"type":"string","default":"UPSC","title":"Exam"}},{"name":"include_rejected","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Include Rejected"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/articles/{article_id}":{"get":{"summary":"Get Article","operationId":"get_article_api_articles__article_id__get","parameters":[{"name":"article_id","in":"path","required":true,"schema":{"type":"string","title":"Article Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/articles/{article_id}/read":{"post":{"summary":"Mark Read","operationId":"mark_read_api_articles__article_id__read_post","parameters":[{"name":"article_id","in":"path","required":true,"schema":{"type":"string","title":"Article Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReadRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/dashboard/{user_id}":{"get":{"summary":"Dashboard","operationId":"dashboard_api_dashboard__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/revision/due/{user_id}":{"get":{"summary":"Revision Due","operationId":"revision_due_api_revision_due__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/revision/submit":{"post":{"summary":"Revision Submit","operationId":"revision_submit_api_revision_submit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RevisionSubmit"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/ingest/run":{"post":{"summary":"Trigger Ingest","operationId":"trigger_ingest_api_ingest_run_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ReadRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"exam_type":{"type":"string","title":"Exam Type","default":"UPSC"}},"type":"object","required":["user_id"],"title":"ReadRequest"},"RevisionSubmit":{"properties":{"user_id":{"type":"string","title":"User Id"},"schedule_ids":{"items":{"type":"string"},"type":"array","title":"Schedule Ids"},"correct_count":{"type":"integer","title":"Correct Count"},"total_count":{"type":"integer","title":"Total Count"}},"type":"object","required":["user_id","schedule_ids","correct_count","total_count"],"title":"RevisionSubmit"},"UserCreate":{"properties":{"name":{"type":"string","title":"Name","default":"Student"},"target_exam":{"type":"string","title":"Target Exam","default":"UPSC"}},"type":"object","title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}